# docs/swagger.yaml
basePath: /
definitions:
...
  handler.LoginRequest:
    properties:
      device_id:
        type: string
      device_name:
        type: string
      email:
        type: string
      password:
        type: string
    required:
    - device_id
    - email
    - password
    type: object
...
  handler.RegisterRequest:
    properties:
      display_name:
        maxLength: 100
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - display_name
    - email
    - password
    - username
    type: object
...
  handler.presenceRequest:
    properties:
      status:
        description: '''online'', ''away'', ''busy'', ''offline'''
        type: string
    required:
    - status
    type: object
  handler.presignRequest:
    properties:
      content_type:
        type: string
      filename:
        type: string
    required:
    - filename
    type: object
  response.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
      success:
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Saybridge Platform Backend API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Saybridge API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login user and generate access & refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Auth failure
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login to account
      tags:
      - Auth
...
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with username, email, password, and display
        name
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Registration successful
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Registration fail (conflict)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
...
  /api/v1/files/presign:
    post:
      consumes:
      - application/json
      description: Generate a presigned URL for uploading a file to S3-compatible
        storage
      parameters:
      - description: File upload details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.presignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Presigned upload and download URLs
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get presigned upload URL
      tags:
      - Files
...
  /api/v1/messages/room/{roomId}:
    get:
      description: Retrieve paginated message history for a room using cursor-based
        pagination
      parameters:
      - description: Room ID (UUID)
        in: path
        name: roomId
        required: true
        type: string
      - default: 50
        description: Number of messages to return
        in: query
        name: limit
        type: integer
      - description: 'Cursor: return messages before this message ID'
        in: query
        name: before_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Message list with pagination cursor
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get message history
      tags:
      - Messages
  /api/v1/messages/thread/{parentId}/replies:
    get:
      description: Retrieve all replies in a message thread
      parameters:
      - description: Parent message ID (UUID)
        in: path
        name: parentId
        required: true
        type: string
      - description: Room ID containing the thread
        in: query
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Thread replies
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get thread replies
      tags:
      - Messages
...
  /api/v1/search/messages:
    get:
      description: Full-text search across messages in accessible rooms
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Filter by room ID
        in: query
        name: room_id
        type: string
      - default: 20
        description: Max results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Search failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search messages
      tags:
      - Search
  /api/v1/search/rooms:
    get:
      description: Search for rooms by name within accessible rooms for the user
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 20
        description: Max results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Matching rooms
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Search failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search rooms
      tags:
      - Search
  /api/v1/search/users:
    get:
      description: Search for users by username or display name within the tenant
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 20
        description: Max results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Matching users
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Search failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search users
      tags:
      - Search
...
  /api/v1/users/presence:
    post:
      consumes:
      - application/json
      description: Set the user's online status (online, away, busy, offline) and
        broadcast to all connected clients
      parameters:
      - description: Presence status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.presenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Presence updated
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Cache error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user presence
      tags:
      - Presence
...
securityDefinitions:
  BearerAuth:
    description: Enter your JWT token with the Bearer prefix, e.g. "Bearer eyJhbGci..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"